IMPORTANT NOTE: Enum change toString() change
Without Object.prototype in the prototype chain, we need toString()
in order to display symbols.
Unfortunately, javascript uses the toString() value if the Symbol is used as an object
index. For example:
       foo[enum] = "some value" is equivalent to foo[enum.toString()] = "some value"
We would prefer when the enum.dbCode to be used as the index value
This also makes it easy to handle serialization by other means.